+++ NpgsqlRestTests/TsClientTests/UploadSseTests.cs
@@ -49,13 +49,13 @@ interface ITsclientTestUploadWithSseRequest {
 }
 
 interface ITsclientTestUploadWithSseResponse {
-      type: string;
-      fileName: string;
-      contentType: string;
-      size: number;
-      success: boolean;
-      status: string;
-      [key: string]: string | number | boolean;
+    type: string;
+    fileName: string;
+    contentType: string;
+    size: number;
+    success: boolean;
+    status: string;
+    [key: string]: string | number | boolean;
 }
 
 
@@ -109,9 +109,9 @@ export async function tsclientTestUploadWithSse(
         }
         xhr.onload = function () {
             if (this.status >= 200 && this.status < 300) {
-                resolve();
+                resolve(JSON.parse(this.responseText) as ITsclientTestUploadWithSseResponse[]);
             } else {
-                resolve({status: this.status, response: undefined!, error: JSON.parse(this.responseText) as {{status: number; title: string; detail?: string | null}}});
+                throw new Error(this.responseText);
             }
         };
         xhr.onerror = function () {